Slight reorganization of UI code
Since we are pivoting to not designing for foldables, the 'compact' etc.
qualifiers for packages and names is no longer relevant.  This reorganizes UI
code to eliminate those qualifiers and reduce confusion.
