+++ native/NativeSampleApps/MobileSyncCompose/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SalesforceMobileSDKAndroid">
         <activity
-            android:name=".MainActivity"
+            android:name=".ContactsActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.SalesforceMobileSDKAndroid">
